% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-variants.R
\name{get_variants}
\alias{get_variants}
\title{Get variants in range}
\usage{
get_variants(con, chromosome, start, end = NULL, isdplyr = FALSE)
}
\arguments{
\item{con}{a database connection object}

\item{chromosome}{the chromosome - must be 1-22, X, or Y}

\item{start}{the first locus to look for a variant}

\item{end}{the last locus to look for a variant (default N/A)}

\item{isdplyr}{logical indicating whther database connection is dplyr (default FALSE)}
}
\value{
A data frame with chromosome, position, reference allele, and alternate allele for each variant within the range (including start and end loci)
}
\description{
Get variants in range
}
\examples{
single_locus <- get_variants(con = con, chromosome = 1, start = 13980)
range <- get_variants(con = con, chromosome = 1, start = 10000, end = 12000)

}
